+++ CHANGELOG.md
@@ -86,8 +86,8 @@
   - fix color
   - add color
     - air-primary, air-primary-success, air-primary-alert, air-primary-copilot, air-primary-warning,
-    - air-secondary, air-secondary-alert, air-secondary-accent, air-secondary-accent-1, air-secondary-accent-2, air-secondary-no-accent,
-    - air-tertiary, air-tertiary-accent, air-tertiary-no-accent
+    - air-secondary, air-secondary-alert, air-secondary-accent, air-secondary-accent-1, air-secondary-accent-2,
+    - air-tertiary
   - @deprecate
     - default, danger, success, warning, primary, secondary, collab, ai
   - add prop inverted